+++ pyproject.toml
@@ -113,7 +113,6 @@ docs = [
     "mkdocs-print-site-plugin >= 2.7.2",
     "mkdocs-exclude >= 1.0.2",
     "pylint >= 4.0.5",
-    "ruff >= 0.9.0",
 ]
 dev = [ # brew install weasyprint
     "qmcpy[docs,test,torch,gpytorch,botorch,umbridge]",
